feat(alerts): add API fn and hook
- Moves alert types to designated file as well.
