+++ src/lib/api/alerts.ts
@@ -1,0 +1,6 @@
+import { alerts } from '@/lib/mocks/alerts';
+
+export async function getAlerts() {
+  return alerts;
+  // return backend.get<Alert[]>('/api/alerts');
+}
