+++ src/lib/api/index.ts
@@ -1,4 +1,6 @@
 export * from './auth';
+export * from './alerts';
+export * from './alerts.d';
 export * from './employees';
 export * from './employees.d';
 export * from './vehicles';
