+++ src/lib/hooks/api/index.ts
@@ -1,2 +1,3 @@
+export * from './alerts';
 export * from './employees';
 export * from './vehicles';
